+++ src/xyz.cpp
@@ -110,7 +110,7 @@ void write_xyz(std::ofstream &file, Atoms &atoms) {
     file << std::endl;
 
     // Element name, position
-    for (int i = 0; i < atoms.nb_atoms(); ++i) {
+    for (size_t i = 0; i < atoms.nb_atoms(); ++i) {
         file << std::setw(2) << "Au" << " "
              << std::setw(10) << atoms.positions.col(i).transpose()
              << std::setw(10) << atoms.velocities.col(i).transpose()
